monitoring/requirements.txt
flask
apscheduler
requests
python-dotenv